buffer: split long lines up
When lines are longer then the window they are on, they get wrapped onto
the next line on the window, but not on the buffer, resulting in
overwriting of wrapped text.

This code checks each added line, and if it's too long, it will split
that line up in the buffer.

For improvements, splits could be made on whitespace only